Added hotkey for line highlighter
Added hotkey for line highlighter.
Refactored line highlighter document for easier function calls. Move lineHighlighter.ts to commands folder
