+++ src/extension.ts
@@ -2,9 +2,9 @@ import * as vscode from "vscode";
 import * as pl from "./pylex";
 import CommandNodeProvider from "./commandNodeProvider";
 import Logger from "./log";
-import { lineHighlighter } from "./lineHighlighter";
 import { installer } from "./pythonManager";
 import path = require("path");
+import {toggleLineHighlight, highlightDeactivate} from "./commands/lineHighlighter"
 
 import {
 	accessCommands,
@@ -13,7 +13,9 @@ import {
 	textCommands,
 	voicetotextCommands,
 	midicommands,
+	lineHighlightercommands
 } from "./commands";
+
 //import { runClient } from "./client";
 
 // Output Logger
@@ -28,8 +30,6 @@ export const rootDir = path.dirname(__filename);
 export function activate(context: vscode.ExtensionContext) {
 	//python packages installer
 	installer();
-	//line highlighter
-	lineHighlighter();
 
 	//runClient(serverModule);
 
@@ -41,6 +41,7 @@ export function activate(context: vscode.ExtensionContext) {
 		navCommands,
 		textCommands,
 		midicommands,
+		lineHighlightercommands
 	].flat(1);
 
 	voicetotextCommands.forEach((command) => {
@@ -67,7 +68,10 @@ export function activate(context: vscode.ExtensionContext) {
 	let hubProvider = new CommandNodeProvider(hubCommands);
 	vscode.window.registerTreeDataProvider("hubActions", hubProvider);
 
+	toggleLineHighlight();
+
 	vscode.window.showInformationMessage("Mind Reader finished loading!");
 }
 
 export function deactivate() {}
+highlightDeactivate();
